Add files via upload
The Snake Game is a classic arcade game that gained immense popularity in the late 1970s and early 1980s. In this timeless and straightforward game, players control a snake that maneuvers around a confined space, typically a bordered rectangular grid. The primary objective is to consume food items, represented by small pellets or dots, scattered throughout the playing field. As the snake devours these items, it grows longer, intensifying the challenge.

The game becomes progressively more difficult as the snake's length increases, requiring players to skillfully navigate without colliding with the walls or the snake's own tail. Collision results in game over. Known for its addictive simplicity and strategic gameplay, the Snake Game has endured through various iterations on different platforms, showcasing its enduring appeal and legacy in the realm of casual gaming.